% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackBill.R
\name{blackBillUI}
\alias{blackBillUI}
\title{按纽生成生成器界面}
\usage{
blackBillUI(
  tabTitle = "黑名单",
  colTitles = c("数据模块", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = blackBillUI_left,
  func_right = blackBillUI_right,
  func_bottom = blackBillUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
blackBillUI()
}
