% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerDueDays.R
\name{customerDueDaysUI}
\alias{customerDueDaysUI}
\title{按纽生成生成器界面}
\usage{
customerDueDaysUI(
  tabTitle = "客户账期规则表",
  colTitles = c("数据模块", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = customerDueDaysUI_left,
  func_right = customerDueDaysUI_right,
  func_bottom = customerDueDaysUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
customerDueDaysUI()
}
